+++ grpcalchemy/server.py
@@ -1,21 +1,52 @@
 import importlib
+import logging
 import time
 from collections import defaultdict
 from concurrent import futures
 from functools import partial
-from typing import Type, Union, Dict, Callable, List, DefaultDict, Any
-import logging
+from typing import Any, Callable, DefaultDict, Dict, List, Type, Union
+
 import grpc
 
-from .blueprint import Blueprint
+from .blueprint import Blueprint, Context
 from .config import default_config
+from .orm import Message
 from .utils import generate_proto_file
 
 _ONE_DAY_IN_SECONDS = 60 * 60 * 24
 
 
-class Server:
-    def __init__(self, config: Union[str, Type] = None):
+class Server(Blueprint):
+    def __init__(
+            self,
+            name: str,
+            file_name: str = '',
+            pre_processes: List[Callable[[Message, Context], Message]] = None,
+            post_processes: List[Callable[[Message, Context], Message]] = None,
+            config: Union[str, Type] = ''):
+        """The Server object implements a base application and acts as the central
+        object. It is passed the name of gRPC Service of the application. Once it is
+        created it will act as a central registry for the gRPC Service.
+
+        Usually you create a :class:`Server` instance in your main module or
+        in the :file:`__init__.py` file of your package like this::
+
+            from grpcalchemy import Server
+            app = Server('server')
+
+        :param str name:
+        :param str file_name:
+        :param List[Callable[[Message, Context], Message]] pre_processes:
+        :param List[Callable[[Message, Context], Message]] post_processes:
+        :param Union[str, Type] config:
+
+        .. versionadded:: 0.2.0
+        """
+        super().__init__(
+            name=name,
+            file_name=file_name,
+            pre_processes=pre_processes,
+            post_processes=post_processes)
 
         #: The configuration dictionary as :class:`Config`.  This behaves
         #: exactly like a regular dictionary but supports additional methods
@@ -47,16 +78,22 @@ class Server:
         self.logger.setLevel(logging.DEBUG)
         self.logger.addHandler(logging.StreamHandler())
 
-        generate_proto_file()
+        self.register_blueprint(self)
 
-    def register(self, bp: Blueprint):
-        grpc_pb2_module = importlib.import_module(f".{bp.file_name}_pb2_grpc",
-                                                  self.config["TEMPLATE_PATH"])
-        getattr(grpc_pb2_module,
-                f"add_{bp.file_name}Servicer_to_server")(bp, self.server)
+    def register_blueprint(self, bp: Blueprint):
+        #: all the gRPC service register in a dictionary by name.
+        #:
+        #: .. versionadded:: 0.2.0
         self.blueprints[bp.name] = bp
 
     def run(self, port: int = 50051, test=False):
+        generate_proto_file()
+        for name, bp in self.blueprints.items():
+            grpc_pb2_module = importlib.import_module(
+                f".{bp.file_name}_pb2_grpc", self.config["TEMPLATE_PATH"])
+            getattr(grpc_pb2_module,
+                    f"add_{bp.file_name}Servicer_to_server")(bp, self.server)
+
         for func in self.listeners["before_server_start"]:
             func(self)